move FuncBuilderAllocations to translator module
This is were it belongs. Also renamed it accordingly.
